i386 target compiles, but does not work

+++ include/kernel/types.h
@@ -9,7 +9,7 @@
 typedef unsigned char  byte;
 typedef unsigned short word;
 typedef unsigned int   dword;
-typedef unsigned long  qword;
+typedef unsigned long long qword;
 
 typedef unsigned char  uchar;
 typedef unsigned short ushort;
